requirements.txt
Django==4.2.18
psycopg2-binary==2.9.9
python-dotenv==1.0.0
djangorestframework==3.14.0
stripe==7.11.0
whitenoise==6.6.0
pillow==10.1.0
django-cors-headers==4.3.1
drf-yasg==1.21.7
python-dateutil==2.8.2
django-environ==0.11.2
gunicorn==21.2.0
